+++ src/Backup.ts
@@ -374,6 +374,8 @@ class Backup {
       "backupVersion",
       "backupPlugins",
       "createSubfolder",
+      "createSubfolder",
+      "exportFormat",
     ];
 
     this.log.verbose("Plugin settings:");
@@ -990,17 +992,20 @@ class Backup {
         }
       } else {
         backupDestination = this.backupBasePath;
-        const oldBackupData = fs
+        const backupData = fs
           .readdirSync(this.activeBackupPath, { withFileTypes: true })
           .map((dirent) => dirent.name);
-        for (const file of oldBackupData) {
+        for (const file of backupData) {
+          let dst = path.join(backupDestination, file);
           try {
-            fs.moveSync(
-              path.join(this.activeBackupPath, file),
-              path.join(backupDestination, file)
-            );
+            fs.moveSync(path.join(this.activeBackupPath, file), dst, {
+              overwrite: true,
+            });
           } catch (e) {
             await this.showError("moveFinishedBackup: " + e.message);
+            this.log.error(
+              path.join(this.activeBackupPath, file) + " => " + dst
+            );
             throw e;
           }
         }
@@ -1021,7 +1026,6 @@ class Backup {
 
   private async clearBackupTarget(backupPath: string) {
     this.log.verbose(`Clear backup target`);
-
     // Remove only files
     const oldBackupData = fs
       .readdirSync(backupPath, { withFileTypes: true })
@@ -1036,23 +1040,30 @@ class Backup {
         try {
           fs.removeSync(path.join(backupPath, file));
         } catch (e) {
-          await this.showError("" + e.message);
+          await this.showError("clearBackupTarget " + e.message);
           throw e;
         }
       }
     }
 
     try {
+      fs.removeSync(path.join(backupPath, "notes"));
+    } catch (e) {
+      await this.showError("clearBackupTarget " + e.message);
+      throw e;
+    }
+
+    try {
       fs.removeSync(path.join(backupPath, "templates"));
     } catch (e) {
-      await this.showError("deleteOldBackupSets" + e.message);
+      await this.showError("clearBackupTarget " + e.message);
       throw e;
     }
 
     try {
       fs.removeSync(path.join(backupPath, "profile"));
     } catch (e) {
-      await this.showError("deleteOldBackupSets" + e.message);
+      await this.showError("clearBackupTarget " + e.message);
       throw e;
     }
   }