+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## not released
 
 - Fix: #47 Suppress repeating error message during automatic execution if the backup path is not accessible
+- Fix: #48 File already exists when a RAW or MD Frontmatter backup with no revisions is made
 
 ## v1.2.0 (2022-11-20)